Suggest to provide credit and link
I like your read me, it has a lot of photos and you provided instructions on how to create this project.

It'd be better to also provide the link to your project, and also show that who did what part of the code. 

This will make the Code Review much easier.